+++ src/main/java/insane96mcp/progressivebosses/mixin/EndCrystalMixin.java
@@ -1,5 +1,6 @@
 package insane96mcp.progressivebosses.mixin;
 
+import insane96mcp.progressivebosses.module.dragon.DragonFeature;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.boss.enderdragon.EndCrystal;
 import org.spongepowered.asm.mixin.Mixin;
@@ -12,7 +13,7 @@ public class EndCrystalMixin {
 
 	@Inject(at = @At("HEAD"), method = "hurt(Lnet/minecraft/world/damagesource/DamageSource;F)Z", cancellable = true)
 	private void hurt(DamageSource source, float amount, CallbackInfoReturnable<Boolean> callback) {
-		//if (CrystalFeature.onDamageFromExplosion(source))
-		//	callback.setReturnValue(false);
+		if (DragonFeature.onCrystalDamagedByExplosion(source))
+			callback.setReturnValue(false);
 	}
 }